Edited intro chapter
The chief change in this chapter has been to replace single quotation marks (') with double quotes (') in all text -- including quite a few of the bits that are hyperlinked to URLs. Unless there is some non-obvious syntax issue I should be aware of for keeping these single quotes, I will be making the same changes throughout the book. A few other changes are to conform to Chicago Manual of Style rules, notably period inside quotation marks at the end of a sentence. Logical? Not really. But as with most syntax rules, sometimes stuff is arbitrary and you just have to work with it. Finally, you'll see a few places where I've broken long sentences into two, so each distinct ideas has its own breathing room, and a few word tinkers just to help a sentence's cadence. ("that" is often helpful for clarity. But not always.)