Edited creating a get request
I haven't found a specific rationale for capitalizing "Outside-in," so I lower-cased it. Please correct me if I've missed some formal origin for the term that warrants the caps.
On line 272, I presume the URL following "curl" in parentheses represents a string to be entered while running a procedure, and shouldn't be formatted as a hyperlink. Correct?