Edited api client
In line 3, the "planeteer" reference is probably meaningful to people who were TV-watching kids in the early 1990s. But it's obscure for any reader who's not in that demographic and didn't grow up on "Captain Planet." This isn't a universal-enough cultural reference, in my opinion, to risk head-scratching by older (or younger) readers.