Rename device token to auth token
`auth_token` more closely reflects the token's new use.

There would normally be a migration to rename `device_token` to `auth_token`,
but in this case we actually want it to appear
like we had always called the field `auth_token`.
So I've updated the existing migrations
and recreated the database to update the schema.