Add bin/setup and bin/deploy
The Rails application is deployed on Heroku from the app in
`example_apps/rails`. It was previously deployed from a separate
`humon-rails` repository, but that was before the app was imported
directly into the book repo.

This adds `bin/setup` and `bin/deploy` scripts based on those generated
by suspenders for deploying the Rails application from this repository.