Automatically define validations from the database
Ensuring validity of data makes sense as a database concern. Most SQL
databases, including Postgres, MySQL, Oracle, and SQLite, include
facility for validating both presence of columns and uniqueness of their
contents (even in scope of other columns) at the INSERT/UPDATE level.

Defining these validations in the form of column constraints and unique
indexes is a best practice in any Rails application, and are supported
as first-class Rails features in migrations, schema statements, etc.
They help us to ensure data integrity at a very low level that can
protect us in cases of shared access to databases (including by other
programs such as psql), the dangers of `validate: false`, and any other
higher-level interactions.

Unfortunately in the case of Rails, we miss out on the nice validation
checks that Rails provides through `#valid?` and indirectly through
`#save`. The error messages we can pass along to our users to help them
resolve issues are features we don't want to miss out on.

With a few lines of introspection on columns and indexes defined in the
database, we can automatically define presence, absence, and uniqueness
validations in the same way that ActiveRecord defines attribute getters
and setters, its `#inspect` string, and other features.

We can also "abuse" database object comments (not to be confused with
SQL language comments) to store messages using syntax inspired by Go's
struct tags. In this case, we're using the presence_message,
absence_message, and uniqueness_message "tags" in column and index
comments to define the nonstandard messages (in this case only the "Can
only RSVP once per event" uniqueness message on event_id and user_id)
that we display to users.

As an aside, I generally recommend that these messages be stored in
translations files instead of in the code or in databases. Translations
are commonly used for allowing messages to be changed for different
locales. This approach also allows for easier [testing] of the messages
and with proper use, a single place to find all user-facing text in an
application. We also avoid describing display logic in models or the
database.

[testing]: <https://robots.thoughtbot.com/better-tests-through-internationalization> "Better Tests Through Internationalization"

We're still able to define conditional validations, validations only
triggered on create or update, and other types of validations such as
inclusion or format in code. These could still be defined in the
databases, most using CONSTRAINTs or TRIGGERs, but they would be more
difficult to programmatically detect as they would use more free-form
definitions. In theory, going back to the tags approach we used for
messaging, we could "tag" columns with validations
(`'validate_inclusion: "[''one'', ''two'', ''three'']"
inclusion_message: "..."'`). I don't perceive as much benefit in this
approach without perhaps a set of migration methods to generate specific
validations in SQL snippets as well as the comments.

A potential downside of this approach is that we no longer see the
validations in code. I don't see this as a big tradeoff, as generally
you also won't see column-backed attributes in a model either. The
precedent has been set that the database drives a lot of functionality
in Rails, and it makes a lot of sense to continue that given the
benefits. Speaking of benefits, the approach of defining the validations
in one place and having the mirrored in others avoids surprising
disparities where you can write data in one layer but not in the other
(Rails validations, no database validations) or when there are useful
errors in one place but not in the other (database validations, no Rails
validations). This can happen when, for example, a null constraint is on
the database but the Rails validation is missing.